Move FiveM entity to separate file
