+++ homeassistant/components/fivem/binary_sensor.py
@@ -10,8 +10,8 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import FiveMEntity, FiveMEntityDescription
 from .const import DOMAIN, NAME_STATUS
+from .entity import FiveMEntity, FiveMEntityDescription
 
 
 @dataclass
